Move legacy app to is own folder
